requirements.txt
Brotli==1.0.9
click==8.0.4
dash==2.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.3
Flask-Compress==1.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.2
Pillow==9.0.1
plotly==5.6.0
six==1.16.0
tenacity==8.0.1
Werkzeug==2.0.3
